UIQM-399 Linked MARC bib field - editable textbox does not support shortcut keys
