+++ src/hooks/useSubfieldNavigation/index.js
@@ -1,2 +1,1 @@
-export * from './useAuthorityLinking';
 export * from './useSubfieldNavigation';
